API Integration with Axios
New Axios client for API requests, used until now for the transactions
listing and for new transactions creation.
Also, now there is a dynamic and an static Money formatter hooks, for
situations where the formatter needs to be dynamic and other for when
we just need a fix value to be masked.